+++ src/components/myParas/MyParas.tsx
@@ -1,7 +1,7 @@
 import { trpc } from "@/client/lib/trpc";
 import React from "react";
-import PersonTable from "./PersonTable";
-import { Para, ParaHeadCell } from "./types/table";
+import PersonTable from "../personTable/PersonTable";
+import { Para, ParaHeadCell } from "../table/table";
 
 const MyParas = () => {
   const utils = trpc.useContext();
