+++ src/pages/paraDashboard.tsx
@@ -1,6 +1,6 @@
 //this will be the mobile-first web view of the para data collection app
 import React from "react";
-import TaskCard from "@/components/para_trials/taskCard";
+import TaskCard from "@/components/taskCard/taskCard";
 import { trpc } from "@/client/lib/trpc";
 // import styles from "@/styles/Paraflow.module.css";
