+++ src/pages/staff/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MyParas from "@/components/case_manager/MyParas";
+import MyParas from "@/components/myParas/MyParas";
 
 function Staff() {
   return <MyParas />;
